+++ test/test_rpc_api.rb
@@ -4,12 +4,12 @@ require_relative 'helper'
 
 class RPCAPITest < Minitest::Test
   class TestAPI < Syntropy::RPCAPI
-    def get(ctx)
+    def get(req)
       @value
     end
 
-    def set!(ctx)
-      @value = ctx.params[:v]
+    def set!(req)
+      @value = req.query[:v]
       true
     end
   end
@@ -20,25 +20,21 @@ class RPCAPITest < Minitest::Test
 
   def test_rpc_api
     req = mock_req(':method' => 'GET', ':path' => '/')
-    ctx = Syntropy::Context.new(req)
-    @app.call(ctx)
+    @app.call(req)
     assert_equal Qeweney::Status::BAD_REQUEST, req.response_status
 
     req = mock_req(':method' => 'GET', ':path' => '/?q=get')
-    ctx = Syntropy::Context.new(req)
-    @app.call(ctx)
+    @app.call(req)
     assert_equal Qeweney::Status::OK, req.response_status
     assert_equal({ status: 'OK', response: nil }, req.response_json)
 
     req = mock_req(':method' => 'POST', ':path' => '/?q=set&v=foo')
-    ctx = Syntropy::Context.new(req)
-    @app.call(ctx)
+    @app.call(req)
     assert_equal Qeweney::Status::OK, req.response_status
     assert_equal({ status: 'OK', response: true }, req.response_json)
 
     req = mock_req(':method' => 'GET', ':path' => '/?q=get')
-    ctx = Syntropy::Context.new(req)
-    @app.call(ctx)
+    @app.call(req)
     assert_equal Qeweney::Status::OK, req.response_status
     assert_equal({ status: 'OK', response: 'foo' }, req.response_json)
   end