+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1 2025-06-17
 
+- Implement routing
 - Implement RPC API controller
 - Implement Context with parameter validation
 - Preliminary version